Added Simulator to project Moved IElevator to correct package for RMI
