+++ src/main/java/at/fhhagenberg/sqe/App.java
@@ -1,58 +1,35 @@
 package at.fhhagenberg.sqe;
 
 import java.io.IOException;
+import java.net.MalformedURLException;
+import java.rmi.Naming;
+import java.rmi.NotBoundException;
 import java.rmi.RemoteException;
 import java.util.Timer;
 import java.util.TimerTask;
 
 import at.fhhagenberg.sqe.model.*;
-import at.fhhagenberg.sqe.model.Elevator.ElevatorDirection;
-import at.fhhagenberg.sqe.model.Elevator.ElevatorDoorStatus;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import org.mockito.Mock;
-import org.mockito.Mockito;
+import sqelevator.IElevator;
 
 /**
  * JavaFX App
  */
 public class App extends Application {
 
-	@Mock
-	ElevatorHardwareManager iElevatorMock;
-
 	private Timer timer;
 	private TimerTask task;
 	private long TIMER_PERIOD = 1000l; // milliseconds
 	
-	public ElevatorHardwareManager getHardwareConnection() throws IllegalArgumentException, HardwareConnectionException, RemoteException {
-		if (iElevatorMock == null) {
-			iElevatorMock = Mockito.mock(ElevatorHardwareManager.class);
-			Mockito.when(iElevatorMock.getFloorNum()).thenReturn(2);
-			Mockito.when(iElevatorMock.getElevatorNum()).thenReturn(1);
-			Mockito.when(iElevatorMock.getElevatorPosition(0)).thenReturn(1, 2,3, 4);
-			Mockito.when(iElevatorMock.getCommittedDirection(0)).thenReturn(ElevatorDirection.Up);
-			Mockito.when(iElevatorMock.getElevatorButton(0, 1)).thenReturn(true);
-			Mockito.when(iElevatorMock.getElevatorButton(0, 0)).thenReturn(false);
-			Mockito.when(iElevatorMock.getElevatorDoorStatus(0)).thenReturn(ElevatorDoorStatus.Closed);
-			Mockito.when(iElevatorMock.getElevatorFloor(0)).thenReturn(0);
-			Mockito.when(iElevatorMock.getFloorButtonDown(0)).thenReturn(true);
-			Mockito.when(iElevatorMock.getFloorButtonDown(1)).thenReturn(false);
-			Mockito.when(iElevatorMock.getFloorButtonUp(0)).thenReturn(false);
-			Mockito.when(iElevatorMock.getFloorButtonUp(1)).thenReturn(true);
-			Mockito.when(iElevatorMock.getServicesFloors(0, 0)).thenReturn(true);
-			Mockito.when(iElevatorMock.getServicesFloors(0, 1)).thenReturn(false);
-			Mockito.when(iElevatorMock.getTarget(0)).thenReturn(0,1);
-			
-			
-			// TODO..
-		}
-		return iElevatorMock;
+	public ElevatorHardwareManager getHardwareConnection() throws IllegalArgumentException, HardwareConnectionException, RemoteException, MalformedURLException, NotBoundException {
+		IElevator controller = (IElevator) Naming.lookup("rmi://localhost/ElevatorSim");    	
+		return new ElevatorHardwareManager(controller);
 	}
 
     @Override
-    public void start(Stage stage) throws IOException, HardwareConnectionException {
+    public void start(Stage stage) throws IOException, HardwareConnectionException, IllegalArgumentException, NotBoundException {
     	//Parent root = FXMLLoader.load(getClass().getResource("/elevator_control_center.fxml"));
     	
 
@@ -87,8 +64,8 @@ public class App extends Application {
         stage.show();   
     }
 
-    public static void main(String[] args) {
-        launch();
+    public static void main(String[] args) {			
+		launch();
     }
     
     @Override
