+++ src/main/java/sqelevator/IElevator.java
@@ -1,4 +1,4 @@
-package at.fhhagenberg.sqe.model;
+package sqelevator;
 
 /**
  * High-level interface to access and control the elevator system.
