+++ src/test/java/at/fhhagenberg/sqe/model/ElevatorHardwareManagerTest.java
@@ -6,6 +6,8 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import sqelevator.IElevator;
+
 import java.rmi.RemoteException;
 
 import static org.junit.jupiter.api.Assertions.*;
